# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Minimal service for injecting controlled HTTP failures (status codes, delays, dynamic response bodies) into dependent services during testing.",
        "title": "time-trial API",
        "contact": {
            "name": "time-trial contributors"
        },
        "version": "1.0"
    },
    "host": "localhost:7777",
    "basePath": "/",
    "paths": {
        "/param-resp": {
            "post": {
                "description": "Configures the schema used to generate response bodies when a 200 is returned by /sabotage or /plan/sabotage. Only statusCode=200 is supported. With no body, clears the active configuration (returns {active:false}).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "param-resp"
                ],
                "summary": "Set param-resp configuration",
                "parameters": [
                    {
                        "description": "Response schema configuration. Omit body to clear.",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.EnvelopeSetParamResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.EnvelopeError"
                        }
                    }
                }
            }
        },
        "/param-resp/config": {
            "get": {
                "description": "Returns the active response-body schema exactly as submitted. Returns 404 if no configuration is active.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "param-resp"
                ],
                "summary": "Get param-resp configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.EnvelopeParamRespConfig"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.EnvelopeError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.ItemConfig": {
            "type": "object",
            "properties": {
                "isColection": {
                    "type": "boolean"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Property"
                    }
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "entities.Property": {
            "type": "object",
            "properties": {
                "isRequired": {
                    "type": "boolean"
                },
                "maxLength": {
                    "type": "integer"
                },
                "minLength": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "propertyFloat": {
                    "$ref": "#/definitions/entities.PropertyFloatConfig"
                },
                "propertyInt": {
                    "$ref": "#/definitions/entities.PropertyIntConfig"
                },
                "propertyString": {
                    "$ref": "#/definitions/entities.PropertyStringConfig"
                },
                "propertyUUID": {
                    "$ref": "#/definitions/entities.PropertyUUIDConfig"
                },
                "type": {
                    "$ref": "#/definitions/entities.PropertyType"
                }
            }
        },
        "entities.PropertyFloatConfig": {
            "type": "object",
            "properties": {
                "floatPrecision": {
                    "type": "integer"
                },
                "isAcceptNegativeValue": {
                    "type": "boolean"
                }
            }
        },
        "entities.PropertyIntConfig": {
            "type": "object",
            "properties": {
                "isAcceptNegativeValue": {
                    "type": "boolean"
                }
            }
        },
        "entities.PropertyStringConfig": {
            "type": "object",
            "properties": {
                "chars": {
                    "type": "string"
                }
            }
        },
        "entities.PropertyType": {
            "type": "string",
            "enum": [
                "string",
                "int",
                "float",
                "uuid",
                "string-funny"
            ],
            "x-enum-varnames": [
                "PropertyTypeString",
                "PropertyTypeInt",
                "PropertyTypeFloat",
                "PropertyTypeUUID",
                "PropertyTypeStringFunny"
            ]
        },
        "entities.PropertyUUIDConfig": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "integer"
                }
            }
        },
        "entities.ResponseConfig": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/entities.ItemConfig"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.EnvelopeError": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.ErrorResponse"
                },
                "duration": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "handlers.EnvelopeParamRespConfig": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entities.ResponseConfig"
                },
                "duration": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "handlers.EnvelopeSetParamResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.SetParamRespResponse"
                },
                "duration": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SetParamRespResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "properties": {
                    "type": "integer"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
    }
}